+++ trixnity-vodozemac/src/jvmMain/kotlin/net/folivo/trixnity/vodozemac/Library.jvm.kt
@@ -1,3 +1,3 @@
 package net.folivo.trixnity.vodozemac
 
-actual val InitHook: () -> Unit = { NativeLoader.ensureLoaded() }
+actual val InitHook: () -> Unit = { JvmLoader.load("vodozemac") }