+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support of OAuth2 login
 - Support for Matrix 1.14, 1.15
 - Support extensible events
+- Check IDB Schema with Tests
 
 ### Changed
 
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add StringWrapper's for more type safety
 - Switch to trixnity-crypto-driver interface instead trixnity-olm
 - Bump Dependencies (Kotlin 2.2.21)
+- Implement IndexedDBMediaStore without juullabs indexeddb
 
 ### Deprecated
 
@@ -41,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed gradle warnings
 - Fixed gradle warnings(even more fixes)
 - Fix joining upgraded room, when not invited to it
+- Fix vodozemac library not found on silicon macs
 
 ### Security